+++ apps/pay/src/utils/constants/wallets/index.ts
@@ -1,1 +1,5 @@
-export const TRON_USDT_CONTRACT: string = "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t";
+export const TRON_CONTRACTS: Record<string, string> = {
+	USDT: "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t",
+	USDD: "TPYmHEhy5n8TCEfYGqW2rPxsghSfzghPDn",
+	USD1: "TPFqcBAaaUMCSVRCqPaQ9QnzKhmuoLR6Rc",
+};