+++ apps/pay/src/views/payerForm/components/steps/stepThree/walletTronConnect/IProps.ts
@@ -1,5 +1,5 @@
 export interface IProps {
 	recipientAddress: string | null;
 	amount: string | null;
-	isUsdtToken: boolean;
+	token: string | null;
 }